refactor(#200): move background handlers
